test(web): Add unit tests for AppComponent
