+++ web/src/tests/app.component.spec.ts
@@ -1,0 +1,23 @@
+import { TestBed } from "@angular/core/testing";
+import { AppComponent } from "../app/app.component";
+import { RouterModule } from "@angular/router";
+
+describe("AppComponent", () => {
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [RouterModule.forRoot([]), AppComponent],
+    }).compileComponents();
+  });
+
+  it("should create the app", () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app).toBeTruthy();
+  });
+
+  it("should have the correct app title", () => {
+    const fixture = TestBed.createComponent(AppComponent);
+    const app = fixture.componentInstance;
+    expect(app.title).toEqual("Productiva Mente");
+  });
+});
